realtimecu: lower strict Whisper deployment memory
Pin the reviewed local Whisper listener and its independent health verifier to GPU int8. This cuts measured ASR residency from roughly 2676 MiB to 1606 MiB while preserving exact authored transcript probes and rejects the former float16 command as deployment drift.\n\nGates: focused shuffle x10 PASS 0.857s; cmd normal PASS 34.385s; cmd shuffle x5 PASS 164.534s (164.96s wall); cmd race PASS 90.253s (93.12s wall); cmd vet PASS 0.29s; bash syntax PASS. The live revalidation benchmark is deliberately deferred until the preserved candidate03 float16 listener is replaced at the next crash-safe service boundary.

+++ cmd/openrealtime/realtimecu_deployment.go
@@ -1077,7 +1077,7 @@ func validateRealtimeCUWhisperProcess(
 	}
 	want := []string{
 		process.Arguments[0], "-I", "-S", "-B", realtimeCUWhisperServiceArgument,
-		"--model", modelPath, "--device", "cuda", "--compute-type", "float16",
+		"--model", modelPath, "--device", "cuda", "--compute-type", "int8",
 		"--language", "en",
 	}
 	for _, root := range expectedDependencyRoots {
@@ -1846,7 +1846,7 @@ func probeRealtimeCUWhisper(
 		payload.Status != "ok" || payload.Model != realtimeCULocalASRModel ||
 		payload.ModelPath != modelPath || payload.MaterializedModelPath != materializedPath ||
 		payload.Device != "cuda" ||
-		payload.ComputeType != "float16" || payload.Language != "en" ||
+		payload.ComputeType != "int8" || payload.Language != "en" ||
 		!reflect.DeepEqual(payload.DependencyRoots, dependencyRoots) ||
 		payload.DependencyDigest != material.DependencyDigest ||
 		payload.Revision != material.Revision || payload.Digest != material.LoadedDigest ||
